+++ src/main/java/org/springframework/samples/petclinic/model/Hotel.java
@@ -1,5 +1,6 @@
 package org.springframework.samples.petclinic.model;
 
+import java.time.LocalDate;
 import java.util.List;
 import java.util.Set;
 
@@ -11,6 +12,8 @@ import javax.persistence.JoinColumn;
 import javax.persistence.OneToMany;
 import javax.persistence.Table;
 import javax.validation.constraints.NotNull;
+
+import org.springframework.format.annotation.DateTimeFormat;
 
 import lombok.Data;
 
@@ -26,6 +29,11 @@ public class Hotel extends BaseEntity {
 	@NotNull
 	@Column(name = "ocupadas")
 	private Integer ocupadas;
+	
+	
+//	@Column(name = "dias_llenos")
+//	@DateTimeFormat(pattern = "yyyy/MM/dd")
+//	private LocalDate diasLlenos;
 
 	// lista de reservas del hotel
 	@OneToMany
