+++ src/main/java/org/springframework/samples/petclinic/web/HotelController.java
@@ -1,18 +1,10 @@
 package org.springframework.samples.petclinic.web;
 
-import java.time.LocalDate;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
-import javax.validation.Valid;
-import javax.websocket.server.PathParam;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.samples.petclinic.model.Booking;
 import org.springframework.samples.petclinic.model.Hotel;
 import org.springframework.samples.petclinic.model.Owner;
-import org.springframework.samples.petclinic.model.Pet;
 import org.springframework.samples.petclinic.model.Review;
 import org.springframework.samples.petclinic.service.BookingService;
 import org.springframework.samples.petclinic.service.HotelService;
@@ -24,10 +16,8 @@ import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
-import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
@@ -89,21 +79,7 @@ public class HotelController {
 
 	}
 	
-	public Integer devolverOwnerId() {
-		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-		Object sesion = auth.getPrincipal();
-		UserDetails us = null;
-		if (sesion instanceof UserDetails) {
-			us = (UserDetails) sesion;
-		}
-		String res = us.getUsername();		
-
-			Owner o = (ownerService.findAllOwners().stream().filter(x -> x.getUser().getUsername().equals(res)))
-					.collect(Collectors.toList()).get(0);
-			Integer ownerId = o.getId();
-		return ownerId;
-
-	}
+	
 
 
 	// LISTADO DE TODAS LAS RESERVAS
@@ -146,7 +122,7 @@ public class HotelController {
 
 	}
 
-	// LISTADO DE RESERVAS DEL OWNER
+	// LISTADO DE RESERVAS DE CUALQUIER OWNER CON ID
 	@GetMapping(path = "/owner/{ownerId}")
 	public String listadoReservasPorOwner(@PathVariable("ownerId") int ownerId, ModelMap modelmap) {
 
@@ -154,116 +130,16 @@ public class HotelController {
 		Iterable<Booking> bookings = bookingService.findBookingByOwnerId(ownerId);
 		// Pone las reservas en el modelmap para mandar a la vista
 		modelmap.addAttribute("bookings", bookings);
+		modelmap.addAttribute("owner", ownerService.findOwnerById(ownerId));
+		
 
 		// Redirige a misReservas.jsp
 		return "hotel/misReservas";
 
 	}
 
-	// CREAR UNA NUEVA RESERVA
-	@GetMapping(path = "/booking/new")
-	public String crearBooking( ModelMap modelmap) {
-		if (ownerService.esOwner()) {
-	Integer ownerId=devolverOwnerId();
 	
-		// Crea una lista con las mascotas para que las muestre en el formulario de
-		// nueva reserva
-		List<Pet> pets = new ArrayList<Pet>();
-		pets = ownerService.findOwnerById(ownerId).getPets();
-		
-		// Tambien obtiene el id de la url, para poder crear una reserva con ese owner
-		// que recibe
-		modelmap.addAttribute("booking", new Booking());
-		modelmap.addAttribute("ownerId", ownerId);
-		modelmap.addAttribute("pets", pets);
 
-		// Redirige al formulario editBooking.jsp
-		return "hotel/editBooking";
-}
-else {
-	modelmap.addAttribute("message", "Para poder crear una reserva, tienes que estar logueado como owner");
-	return "welcome";
-}	
-
-	}
-
-	// CREAR UNA NUEVA RESERVA Falta hacerlo bien
-	@PostMapping(path = "/save/{ownerId}")
-	public String guardarBooking(@PathParam("pet") Integer pet, @PathVariable("ownerId") Integer ownerId,
-			ModelMap modelmap) {
-
-		Booking reservaAutogenerada = new Booking();
-		reservaAutogenerada.setEndDate(LocalDate.of(2020, 10, 30));
-		reservaAutogenerada.setStartDate(LocalDate.of(2020, 10, 30));
-
-		reservaAutogenerada.setOwner(ownerService.findOwnerById(ownerId));
-		reservaAutogenerada.setPet(petService.findPetById(pet));
-
-		bookingService.save(reservaAutogenerada);
-		modelmap.addAttribute("message", "Booking creado con éxito!");
-
-		modelmap.clear();
-		String vista = listadoReservasPorOwner(ownerId, modelmap);
-		return vista;
-
-	}
-
-	// BORRAR UNA RESERVA
-	@GetMapping(path = "/delete/{bookingId}/{ownerId}")
-	public String borrarBooking(@PathVariable("bookingId") int bookingId, @PathVariable("ownerId") int ownerId,
-			ModelMap modelmap) {
-		Optional<Booking> booking = bookingService.findBookingById(bookingId);
-
-		// Si la reserva está la borra, si no, te redirecciona a la lista de reservas
-		if (booking.isPresent()) {
-			bookingService.delete(booking.get());
-			modelmap.addAttribute("message", "Booking borrado con éxito!");
-
-		} else {
-			modelmap.addAttribute("message", "Booking no encontrado");
-		}
-
-		// Cuando acaba el metodo, te redirecciona a la lista de reservas del owner
-		return listadoReservasPorOwner(ownerId, modelmap);
-	}
-
-	// nueva reseña al hotel
-	@GetMapping(path = "/review")
-	public String crearReviewHotel(ModelMap modelmap) {
-
-		String vista = listadoReservas(modelmap);
-
-		if (ownerService.esOwner()) {
-
-			// Manda una review vacia al formulario junto al owner, para que se rellene y se
-			// mande al metodo "/save"
-
-			devolverOwner(modelmap);
-			modelmap.addAttribute("review", new Review());
-			vista = "reviews/newReview";
-
-		} else {
-			modelmap.clear();
-			modelmap.addAttribute("message", "Solo los owners pueden hacer reviews del hotel");
-			vista = listadoReservas(modelmap);
-		}
-		return vista;
-
-	}
-
-	// nueva reseña al hotel
-	@PostMapping(path = "/saveReview/{ownerName}")
-	public String guardarReview(@Valid Review review, BindingResult result, ModelMap modelmap,
-			@PathVariable("ownerName") String ownerName) {
-
-		review.setOwnerName(ownerName);
-		// Obtiene la reseña del formulario y la guarda en la bd
-		reviewService.save(review);
-		modelmap.addAttribute("message", "Review creada con éxito!");
-
-		// Cuando acaba, redirecciona a la lista de reservas, donde esta la review
-		return listadoReservas(modelmap);
-
-	}
+	
 
 }
